fixup! Add GitGutterSettings class that encapsulates settings (#230 related)
initialize settings once as global and use as global
