+++ git_gutter_popup.py
@@ -11,16 +11,16 @@ except:
     _MDPOPUPS_INSTALLED = False
 
 try:
-    from .git_gutter_settings import GitGutterSettings
+    from .git_gutter_settings import settings
     from .view_collection import ViewCollection
 except (ImportError, ValueError):
-    from git_gutter_settings import GitGutterSettings
+    from git_gutter_settings import settings
     from view_collection import ViewCollection
 
 _MD_POPUPS_USE_WRAPPER_CLASS = int(sublime.version()) >= 3119
 
 
-def show_diff_popup(view, point, settings, flags=0):
+def show_diff_popup(view, point, flags=0):
     if not _MDPOPUPS_INSTALLED:
         return
 
@@ -173,11 +173,6 @@ class GitGutterReplaceTextCommand(sublime_plugin.TextCommand):
 
 
 class GitGutterDiffPopupCommand(sublime_plugin.WindowCommand):
-    def __init__(self, *args, **kwargs):
-        sublime_plugin.WindowCommand.__init__(self, *args, **kwargs)
-        self.settings = GitGutterSettings()
-        self.settings.load_settings()
-
     def is_enabled(self):
         if not _MDPOPUPS_INSTALLED:
             return False
@@ -188,4 +183,4 @@ class GitGutterDiffPopupCommand(sublime_plugin.WindowCommand):
         if len(view.sel()) == 0:
             return
         point = view.sel()[0].b
-        show_diff_popup(view, point, self.settings, flags=0)
+        show_diff_popup(view, point, flags=0)
